pyproject.toml
[project]
name = "gitwarden"
version = "0.1.0"
description = "Manage multiple git repositories seemlessly(ish)."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click",
    "gitpython",
    "pydantic",
    "pytest",
    "pytest-cov",
    "python-gitlab",
    "rich",
    "rich-click",
]

[project.scripts]
gitwarden = "gitwarden.cli:cli"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
# Target modern Python
target-version = "py310"
line-length = 120
fix = true
show-fixes = true

# Exclude common non-source directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    ".mypy_cache",
    ".pytest_cache",
]

[tool.ruff.lint]
# Enable a strict, opinionated rule set
select = [
    "F",     # Pyflakes (unused imports, undefined names)
    "E",     # Pycodestyle errors
    "W",     # Pycodestyle warnings
    "I",     # Import sorting (isort)
    "N",     # PEP8 naming
    "B",     # Bugbear
    "C4",    # Comprehension optimisations
    "UP",    # PyUpgrade
    "SIM",   # Code simplifications
    "ARG",   # Unused function arguments
    "DTZ",   # Datetime timezone issues
    "RUF",   # Ruff-specific rules
    "ANN",   # Type annotations
    "D",     # Docstrings
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
# Fully Black-compatible formatting
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
